TBR v23: Update package doc for all packages.
The goal is to make the summary godoc page look presentable and
consistent.  The doc for the types and funcs in each package
still needs work; that'll happen later.

FYI I'm using "defines" for packages that mainly contain VDL or
APIs.  I'm using "implements" for packages that actually
implement a piece of functionality.  I'll use this convention
throughout our godoc.

Change-Id: Ibddbdb053206dd659b3b71a2efbc41fc31bd623f