v23: Introduce client.Call that retries application level errors.
If a server returns a retryable verror, client.Call will retry the
call.
client.Call is a synchronous substitute for client.StartCall/call.Finish.

I had to add the method to the Client interface instead of directly in the
v23/rpc repository so that the implementation can be in x/ref/profile/...

MultiPart: 1/2

Change-Id: I852121c063684a2f13d633c7cb772a1f4b52bdd1