"v23/security": BlessingStore.Default
This CL changes the semantics of BlessingStore.Default() so that
it always returns the value set by the most recent SetDefault call,
or empty Blessings if no such call has been made.

Prior to this CL, Default returned the result of ForCall if no
SetDefault call had been made. Since we don't have any particular
reason to support this special-case semantics, this CL gets rid
of it in order to simplify the implementation.

Change-Id: Ibb8a0491a03019efd1314d72db23ee7225fc5d75