v23/verror: chained stack traces for v23.New(newerr, err)
A common pattern for verror usage is the following:

func f() error {
    if err := callA(i); err != nil {
        return verror.New(errNew, ctx, i, err)
    }
    return nil
}

This embeds the error message for the error returned
by CallA into the new error created by verror.New
according to the format string created for errNew.
Unfortunately, the stack trace for error returned
by callA is lost and it's not possible to locate
the code that generated the original error. This CL
addresses this problem by chaining the stack trace
of the first verror.E to occur in a parameter list
to the stack trace of the newly created error, so that:

if err := f(); err != nil {
    fmt.Println(verror.DebugString(err))
}

will print a stack trace that extends into callA and
shows the location of the original creation of the
error returned by callA.

This is obviously extremely useful for debugging and
it also makes it much easier to cleanup error messages
since it's now possible to quickly see where they
are generated.

Change-Id: I27b7a6c97fd8e759060637b8638cb97c7b6d8e28