v23: Change context mutators from SetFoo to WithFoo.
The motivation for this change is consistency, so our API becomes
easier to use.

Before this CL we were using the WithFoo style in the context
package, but the SetFoo style elsewhere, sometimes with an extra
"Context" thrown in for kicks.  After this CL we use WithFoo (and
GetFoo) everywhere.

MultiPart: 1/4
Change-Id: I9c2228183a99940848afa5619362814a6a4e5acd