syncbase: query_[checker|parser]: collect key prexies, compute like regexes.
Key prefixes need to be collected from key expressions in the where
clause as filtering by keys is performed first (because it is most efficient).
Furthermore, like clause string literals need to be checked for errors and
converted to regular expressions (for evaluation during query execution).
Both of the above are performed during semantic checking--since errors
need to be reported then.

Change-Id: I2a43e11a7699e1a57ba8ba6b66952cfd3f12cbd7