v.io/v23/options: Exposing IsLeaf as a RPCServerOpt similar to how ServesMountTable is already an option.
Prior to this, isLeaf was only set if server.Serve() was used. It
was always false if server.ServerDisptacher() was used instead.
WSPR needs to use ServeDispatcher() for all servers since it simply
proxies to JS but what it is serving might be a Leaf regardless.
Similar cases might exist where one needs to use ServeDispacther but is
expected to serve leafs, hence the need to expose it as a Server options

MultiPart: 3/3
Change-Id: Ife0d2353dadaacd2a52ade4e3d0e978cc2ec5591