vom: Fix vom.Dumper status output.
The root problem was reported by Shyam: "vom dump" was reporting
4 errors at the end of every successful decoding; 2 flush and 2
close errors.

The duplication of the flush and close errors was due to
writeStatus being called incorrectly when the peek of the version
byte fails.  Having them appear at all was due to the dump status
not being reset in certain cases.

In addition to fixing the bugs above, I've also changed the code
to allow the optional version byte before *any* message.  This is
more convenient for "vom dump" when cutting-and-pasting hex
strings into the terminal.

Change-Id: Iabf74cec961ec0a566d8107b018d2aa799560077