syncbase: query: add functions to parser only.
The parser now recognizes functions as operands in the where clause.
It does NOT YET recognize functions in the select clause.
Also, the checker currently rejects the functions as not yet implemented.
I want to close out this work--and check it in--before I address a few kash
code reviews of past changes.

Change-Id: If866f38e0e5ed176e279795e4fedceb555d88a52