conventions: Add a directory for conventions we'ld like developers to use.
We start by providing a routine that will return the possible user IDs
of a caller.  User ID isn't really defined anywhere else in veyron.
However, we need something less malleable than BlessingName to perform
tasks like resource accounting.  Here we are using the convetion that
the first three components of the Blessing are the user id; the identity
provider, a single character type, and the actors name. An example is
"dev.v.io/u/p@google.com" which is the user p@google.com as authenticated
by the identity provider dev.v.io.

Change-Id: I38848986f72222560dc63e2acb69c4fe4164cce6