v23/logging: introduce a v23 api for logging and embedd it in context.T
This is a step towards moving logging to be context based so that
we can control logging at a more granular logging - e.g. to create
and/or encrypt logs separately for each client principal that interacts
with the server.

context.T implements v23/logging.Logger directly and hence can
be used for logging - e.g. ctx.Infof, or ctx.VI(2).Infof.

As part of this change, the interfaces defined in x/lib/vlog are
removed and defined in v23/logging instead. x/lib/vlog is an
implementation of this interface. In addition, x/ref/lib/logger
provides a wrapper for x/lib/vlog so that the runtime does not
need to import/use vlog directly.

MultiPart: 1/3
Change-Id: I1e8ebb3c39aa2c9961f6ae2f45c3ab07c31beb66