vdl: Fix bug in byte value assignment.
The bug occurs when typedCopy is called to copy a byte value;
that happens in various places, including map and set key
assignment.

The typedCopy call first creates a vdl.Value{Byte, nil}, and
calls Assign.  For all other kinds of values the Assign works as
expected.  But AssignByte has special-case logic to deal with the
two representations of bytes, both as standalone values and as
elements in an array.  That logic didn't handle the rep==nil
case, causing the bug.

The fix is to simply handle the rep==nil case just like the
standalone uint64 case.  All the rest of this CL is additional
tests.

This bug was identified and reported by John Kline.

Change-Id: Icecce61bd78f5296c776ebb946c6d01fcbe54113