rpc: New Glob/GlobChildren interfaces
This change adds new Glob and GlobChildren interfaces that are more
in line with the other RPC interfaces. In particular, both Glob and
GlobChildren now take a GlobServerCall argument that implements the
SendStream method. In addition, Glob now takes a *glob.Glob argument
instead of a pattern string, and GlobChildren takes a *glob.Element
argument.

The old interfaces are preserved for backward compatibility, but will be
deprecated soon.

Partly addresses v.io/i/532 and v.io/i/533

MultiPart: 1/3
Change-Id: Ic1299b81eeb763e149e2c25492d920e9ee2a921e