security: Cache signature verifications for Discharges.
Resolves vanadium/issues#625

Some applications (such as syncbase when scanning through rows, or
possibly Glob implementations) need to authorize each result that is
returned to the user. Towards this end, then end up calling:
security.RemoteBlessingNames
multiple times.

Each call to RemoteBlessingNames validates all caveats on the blessings.
For third-party caveats, this validation includes a signature
verification, which can take ~3ms of CPU. By caching the signature
verifications, the goal is to significantly speed up this process.

This is similar in spirit to the work done to speed up verification of
the certificate chains that back blessings in vanadium/issues#588

An alternative would be to have applications cache the result of
RemoteBlessingNames and use that instead of calling it each time. This
might be a fine idea, but I believe the caching in this commit is still
beneficial since it transcends boundaries of "ctx" and security.Calls
(if the same set of blessings/discharges are presented in multiple RPCs,
this caching still helps with CPU savings).

For discharges, this caching results in a ~83% reduction in CPU cycles
as demonstrated by these benchmarks:
BenchmarkPublicKeyDischargeVerification_Cached 3000    479052 ns/op
BenchmarkPublicKeyDischargeVerification	        500   2721848 ns/op

The 479ns is still pretty high for a Caveat.Validate call. I traced this
to vom-decoding, which is still happening on every call to
Caveat.Validate. Arguably, the cache could be moved to this higher level
as well.

Also, to consider: Whether the signature caching should be moved to
signature.go so that the same caching applies for all calls to
Signature.Verify instead of having to maintain caches for each type of
signature verification.

Change-Id: I70ec8f9ad050bcceed69b9af51b2b60d2e8117cb