"security": Signer must also sign its own public key
It is considered good crypto hygiene to always include
the public key of the signer in all messages signed by
it. This CL changes the behavior of the ECDSA signer so
that it adheres to this.

This making this change directly would be backward imcompatible,
this CL introduces new signing purposes under which the signer
always includes its public key in the message being signed.

All signing operations performed by Principal objects still
use old purpose to be backwards compatible. In a few weeks,
once most of the binaries have been updated, we will update
the Principal object so that it uses the new purposes.

Addresses vanadium/issues#619

MultiPart: 1/2

Change-Id: Idee007db48143f90be60e387a65b3b1d29f27a5d