v23/flow: Add flow errors vdl to public api.
The flow manager will have to return errors of this type from all
of its methods.

After this CL I will add the ability to skip the automatic vdl generation
of {1:}{2:} in the vdl error format, which is desired by these errors, since
these errors are all expected to be used as arguments to higher level
errors.

Change-Id: Ia66e6a5f79347b08786d804653d500355b5b03bd