"security": Signing Blessings
While signing data at rest, a principal oftens wants to attach
one of its blessings to the data blob in order to establish
authorship of the blob. Use-cases involving signing data at rest
include syncing updated via syncbase, and publishing application
binaries.

Since blessings can carry caveats and the validity of caveats
can vary accross principals and contexts, not all blessings
are suitable to attach as "identities" to data being signed at
rest.

See this doc for a more elaborate argument:
https://docs.google.com/a/google.com/document/d/18ahgURFLK81Nji4YtjkI6mZmxUj0YE8nB8dYBSdikQU/edit?usp=sharing

This CL introduces the concept of "signing blessings" which are
blessings that can only carry caveats that are universally
understandable and validatable. As a result of this restriction,
the names of a signing blessing can be evaluated in all contexts.

In particular, this CL adds two functions:

- SigningBlessings: This function takes a blessings object and
  returns one with just those chains from the provided blessings
  object that have universally understandable and validatable
  caveats.
- SigningBlessingNames: This function takes a Blessings object and
  determines the set of names encapsulated in the object that are
  valid at the time when the function is called.

Currently, we only allow Expiry caveats on signing blessings. In
the future we would also like allow revocation caveats.

MultiPart: 1/2

Change-Id: If03172c373d1e996807fa0896cbe1832525705c9