"security": Add Blessings.CouldHaveNames
This CL adds a 'CouldHaveNames' method to the Blessings type that
checks whether a given set of names is a subset of the names
encapsulated in the blessings object.

This method would be used by syncbase in order to check whether
the blessing names associated with a data blob are valid w.r.t.
the blessings object for the blob.

Change-Id: I7b334383709c0f79b57e6011a42d782c0cafbfd5