"v23/security": Addressing a minor technical nit on how certificate chain digests are computed
Currently the digest of a certificate chain is computed by the
cert.chainedDigest(hashfn, chain) where
-- cert is the final certificate in the chain
-- hashfn is a hash function appropriate for 'cert'
-- chain is the digest of the certificate chain upto the last but one
certificate.

The returns hash is defined as
hashfn(append(chain, hashfn(cert)))

Note that we are appending 'chain' and 'hashfn(cert)'.

According to our technique for hashing tuples, we should always
map each tuple component to a fixed length byte slice and then
concatenate these byte slices.

Above, we violate this technique if 'chain' and hashfn(cert) are
of diferrent lengths. This would happen when the last certificate
in 'chain' and 'cert' have public keys with different number
of bits. Such chains are supported by our framework.

Therefore to stick to our technique, I propose changing the
'chainedDigest' method to return
hashfn(append(hashfn(chain), hashfn(cert)))

This is a super nit :-) I was trying to carefully understand
how chainedDigest worked and I noticed this. I don't think
it is possible to explot this, however there is no harm in fixing
it in the interest of being consistent throughout.

There is no noticeable change in the benchmarks for certificate
chain verification.

Change-Id: Iee45148685f63659ed69916518ecd552e7f490b3