vom: Speed up single-shot vom.Decode via a TypeDecoder cache.
The single-shot vom.Decode is very expensive, since all type
messages are decoded for every encoded value.  This is
particularly bad for syncbase, which uses vom.{En,De}code to read
and write user-supplied values.

There are many ways to speed up vom encoding and decoding; this
is the simplest way to get some quick wins in the common case.
The idea is to maintain a global TypeDecoder cache keyed by the
bytes of the type messages.  It's relatively cheap to take a
first-pass through the bytes passed to vom.Decode, determining
the starting point of the value message, to create the cache key.
If we have a TypeDecoder in the cache we can just use it;
otherwise we run through the regular decoding codepath, and add
the TypeDecoder to our cache for future use.

Change-Id: Iae7841ce6280c6191524b92fd80ac76fe5d315e0