syncbase: syncQL: add Split and Len functions.
Split: splits a string into a list of strings based on a separator.
       Empty separator splits on UTF-8 sequences.
       Just calls string.Split.
       ex: Split("abc.def", ".")

Len: returns the length of a string or the number of entries in
     arrays, lists, maps and sets.
     returns 0 for nil
     returns error (resolves to nil) for all other types
     ex: Len("abc")

In addition to the above, the Type() function collapsed by a couple of lines
and the db argument is now passed into functions (for the context
in creating verrors).

MultiPart: 1/2

Change-Id: Icead1bffbb785334b9015562d611e07c9996c7a5