Make query_checker tests pass in go1.5.
Go 1.5 has moved to storing time zone data in a zip file. The error messages
returned have also changed, causing the query_checker tests to break under 1.5.

This CL addresses the issue by allowing tests to return one of a range of
errors. If the test specifies an expectation of vErrorContaining, the error
returned by the code-under-test will match if:

* its ID matches the expected error
* its error message contains the text of the expected error

This is more lenient than the previous test. Also, the offset parameter is not
checked.

Change-Id: I6bb264a9b171783380fa9a9d0e0ee635cc858534