syncbase: syncql: limit queries to 10K in length.
Syncbaase is (or is about to be) announced.  I don't see any
limits to arg sizes in the client libraries at present, but I
figured, in any event, I should put a limit in the parser.

This change limits the exec query argument to 10K.  Since
go grows the stack, it could be much, much larger; but 10K
seems to be a reasonable proposal for query size (until we
hear further from users).

Change-Id: I8fef89556971c1f1c5a8b15b5a5ac1bbb41eb98a