security: Use bytes.Buffer instead of string appends to reduce allocations.
Following up with a review comment on a previous patch, this one
uses a bytes.Buffer instead of the "+=" operator on strings to
join names. This reduces the number of memory allocations and in
theory should improve performance. A micro-benchmark comparing
the two shows that joining two strings with += vs using bytes.Buffer -
bytes.Buffer is 100ns/op faster

This doesn't show up on other benchmarks in the security package
because the noise there is >100ns/op and the benchmarks use
a small number of certificates. However, these benchmarks show
no appreciable degradation by switching to bytes.Buffer so I feel
like patching this in anyway.

Change-Id: I1ee771a3f4821de5809463668196c3d36039fd16