syncbase: switch to /$/ hierarchy separator scheme
This change is needed so that blessing strings can be
embedded in component names (e.g. app names) without
escaping the slashes. This change could be avoided if
blessing strings were to use any separator other than
'/' - in that case we'd most likely opt to simply
escape slashes.

As part of this change, we can relax the naming on
components (app, db, table, row) to allow almost any
valid UTF-8 character sequence. The restrictions that
remain are as follows:

- As with blessing strings, we disallow $ components
  and @@ strings.
- We disallow leading double underscore (__) in slash-
  separated components since it's used for reserved
  names in system RPCs, e.g. /__debug RPCs.
- To protect our sanity, we disallow empty slash-
  separated components. (Otherwise, we have to be
  really careful everywhere with naming.Join and
  such.)
- We reserve the nil character (\x00).

These restrictions can be further loosened if we see
value in doing so.

MultiPart: 1/5
Change-Id: I44b462b71ad56e8c1378eeb895399fb43a114772