security, options: Symmetric client and server authorization.
Prior to this patch, servers used a security.Authorizer to encapsulate
their policy for authorizing clients. Clients on the other hand used
specially defined options (options.SkipServerEndpointAuthorization,
options.ServerPublicKey, options.AllowedServersPolicy) to define
their policy for server authorization.

This seems to have many downsides which I'm not quite sure why we ended
up with. For example:
- The client had a limited, non-extendable set of policies to apply to
  the server
- The API was uglier - authorization policy of the server was nicely
  encapsulated in an interface but the reverse was encapsulated in
  thre special types which had to be passed to StartCall.
- options.SkipServerEndpointAuthorization applied to both the end
  server and intermediate mounttables while the other options
  only applied to the end server.

This patch changes things so that security.Authorizer should be
the only encapsulation of a policy and it can be provided to
rpc.Client.StartCall to be applied to a server. The option-based
policies are now written out as security.Authorizer implementations.

The policy used to authorize intermediate nameservers (mounttables) is
explicitly called out.

This will be particularly useful in places where the client/server are
just an artefact of which process reached out to which first (for
example, syncbase processes calling GetDeltas on each other).  Now, the
same policy can be applied irrespective of which end is sending the
request.

MultiPart: 1/6
Change-Id: I940fc52e67dfc7a01bb691f974672e3ac132e8b7