security: Make AddToRoots a function instead of a method on the Principal interface.
AddToRoots is really a convenience function. By removing it from the
interface, other Principal implementations are simplified as they
have one less method to worry about.

The larger motivation for this is that we plan to have a way to create
a Principal that is the same as another one, except that it uses a
different BlessingStore and BlessingRoots. For example, one may do this
when they get blessings granted in an RPC and want to dump it in their
store for related RPCs but not affect the original blessing store.
Or one may do this when creating a server that uses a different blessing
than the original BlessingStore.Default.

By moving AddToRoots out of the interface, these "derived" Principals
will be much easier to manage.

MultiPart: 1/7
Change-Id: I410b701ab22182435bd8bdf54eeaaa61a943163f