v23: Ensure endpoint used for StartCall is used for conn.Dial.
Scenario:
User uses a endpoint without a blessing pattern to remote end, thus
making an unauthorized connection to the remote end.
Then client then decides to specify a endpoint with a blessing pattern.
We need to ensure that this new endpoint with the blessing pattern is
the one used to authorize the server during conn.Dial.

This means that flows on the same Conn may have a different RemoteEndpoints,
thus the API changes.

Change-Id: I153ef972b2f61397c08a10cd3f7f9b83876a78df
MultiPart: 1/2