query: Move query packages out of syncbase.
Query is indepenent of syncbase and will be used by other
packages (starting with discovery).

This change moves query out from under syncbase to be a
standalone set of packages.

As part of the many moves query has made, some packages got
combined that shouldn't have.

The query packages implement Vanadium's query capabilities.

syncql:     This package is used by client's of a Vanadium
            component that supports queries.
            It includes:
                - ResultStream (which is the product of
                  executing a query).
                - The error messages that could be returned
                  from performing a query.
                - A function to parse an error to get the
                  offset into a query that caused the error.
            syncql does does not include the Exec function
            as that must be provided by the component that
            supports queries.

engine:     This is the query engine.  A component that
            supports queries calls the Exec function in this
            package.

datasource: This package contains the interfaces that a
            Vanadium component must implement in order to
            use the query engine.

internal:   This is internal code which implements queries.
            Its sole client is the engine package.

This is a multipart change.

MultiPart: 1/2

Change-Id: I874433492e90a2449bc4c264dfb251c452d56038