query: bugfix for unterminated literals
The parser was not honoring scanner errors because some golang
text/scanner errors are actually valid strings in syncQL.  This
presents a problem for unterminated literals.

The parser now honors all scanner errors except for strings.  In
the case of string, it checks that the string was terminated.  This
isn't quite correct as the terminating double quote could have been
escaped, but it's much better than it was.

The solution is what has been planned, to write a quick and dirty scanner
which is correct for the language.  I'll need to get to that sooner
rather than later.

Change-Id: I1222a9b0a66b90c5d24ac428e42b6ac15cba2091