TBR v23/vom: Return current buf length to fix decoder.Ignore.
This fixes TestDatabaseCreateNameValidation.
Specifically, decoder.Ignore was failing when the server was reading
client arguments. This means that the server would sometimes close
the connection due to the lookup error from the test, BEFORE the client
had finished encoding arguments. This made the client sometimes get
an argument encoding error and other times get the actual lookup
error from the server.

Change-Id: Ia60a1a484a9897e3d3b95d8050227c1c0b0c4895