syncQL: add escape clause to select statement
The query language was attempting to use backslash as the escape
character for wildcards when using the like expression.
Backslash was a bad choice because backslash is used to escape
other characters in string literals.

This change introduces the escape clause to specify an escape
character when needed.

For example:
To find all rows where v.Name begins with the literal '%' followed
by anything, one would write:

select k, v from Customers where v.Name like "#%%" escape '#'

In the avove example, the first % is escaped, so it is treated
as a literal.  The second percent is the wildcard.

MultiPart: 1/2
Change-Id: I0b3a661328476edca5df05a74c743535b4d7f96a