services/syncbase/nosql:  Add "IsServer bool" to SyncgroupMemberInfo struct
This changes adds an IsServer member to the SyncgroupMemberInfo
struct, to allow clients to specify whether a particular member
of a syncgroup is a "server".

A "server" in a syncgroup is a device that is assumed to have
higher-than average storage, computing power, bandwidth,
availability, and access-control rights within the syncgroup.
For many syncgroups, it will be a server in the cloud.  For some,
particularly some with major privacy constraints, it may be a
workstation in the syncgroup administrator's home.

Normally, a "server" has permission at least to see all data in
the syncgroup, and has enough storage to store everything in the
syncgroup, including all blobs referenced by the syncgroup.  It
is expected to be a suitable (and preferential) target for other
devices to sync with if they are connected to the internet.
Thus, use of servers forms an informal star pattern for syncing,
and thus tending to reduce the expected "sync-diameter" of the
syncgroup from logarithmic to closer to constant.

It is also expected to act as a "magnet" for blob ownership.  It
is willing to receive blobs and take blob ownership from any
other member.  It is expected to store blobs reliably enough that
no non-server members need store blobs to avoid data loss.

			-----

The changes in the tests in syncbase/nosql merely to use explicit
field names in initializations of the SyncgroupMemberInfo struct,
so that additional fields (like IsServer) may be added without
changing initializers.

MultiPart: 1/2
Change-Id: Idc8ead02e9ba14b6cc54ad9fa66fc73d721dc510